Widen the eraser area (#113)

+++ src/components/Editor/DrawingBoard/Canvas/Worker/EraserWorker.ts
@@ -1,5 +1,5 @@
 import { TimeTicket } from 'yorkie-js-sdk';
-import { Root, Point } from 'features/docSlices';
+import { Root, Point, Box } from 'features/docSlices';
 import { ToolType } from 'features/boardSlices';
 import { createEraserLine, fixEraserPoint } from '../line';
 import Worker from './Worker';
@@ -14,6 +14,8 @@ class EraserWorker implements Worker {
   private emit: Function;
 
   private createID?: TimeTicket;
+
+  private eraserBoxSize: number = 24;
 
   constructor(update: Function, emit: Function) {
     this.update = update;
@@ -49,13 +51,30 @@ class EraserWorker implements Worker {
 
         const findAndRemoveShape = (point1: Point, point2: Point) => {
           for (const shape of root.shapes) {
+            if (shape.type === 'eraser') {
+              continue;
+            }
+
             if (isSelectable(shape)) {
               if (isInnerBox(shape.box, point2)) {
                 root.shapes.deleteByID(shape.getID());
               }
             } else {
               for (let i = 1; i < shape.points.length; i += 1) {
-                const result = checkLineIntersection(point1, point2, shape.points[i - 1], shape.points[i]);
+                const shapePoint1 = shape.points[i - 1];
+                const shapePoint2 = shape.points[i];
+
+                if (isInnerBox(this.eraserBox(point2), shapePoint2)) {
+                  root.shapes.deleteByID(shape.getID());
+                  break;
+                }
+
+                if (isInnerBox(this.eraserBox(point1), shapePoint2)) {
+                  root.shapes.deleteByID(shape.getID());
+                  break;
+                }
+
+                const result = checkLineIntersection(point1, point2, shapePoint1, shapePoint2);
                 if (result.onLine1 && result.onLine2) {
                   root.shapes.deleteByID(shape.getID());
                   break;
@@ -93,6 +112,15 @@ class EraserWorker implements Worker {
       root.shapes.deleteByID(this.createID);
       this.emit('renderAll', root.shapes);
     });
+  }
+
+  eraserBox(point: Point): Box {
+    const x = point.x - this.eraserBoxSize / 2;
+    const y = point.y - this.eraserBoxSize / 2;
+    const width = this.eraserBoxSize;
+    const height = this.eraserBoxSize;
+
+    return { y, x, width, height };
   }
 }
